opa-eac/api/mstr/requirements.txt
boto3
pexpect
mstrio-py==10.11.1
requests
pymysql
pytest-shutil
configparser
psycopg2-binary
pyyaml